First release towards the new MoWGLI exportation module:
Modifications outside the XML module:
1) unsharing function added to Coq's kernel
2) interface of Evarutil extended
3) Makefile (and .depend) changed

Modification inside the XML module:
The exportations is now performed in several phases. This leads to
several modules:
Acic: definition of the internal data-structures isomorph to the DTDs
Acic2Xml: mapping from the internal data-structure for objects to a XML stream
Cic2acic: mapping from constr to the internal data-structures isomorph to the
          DTDs
DoubleTypeInference: Coscoy's style type-inference
Proof2aproof: mapping from the proof-trees of Coq to the lambda-term that
              is the corresponding proof. Cross-linking information is
              retained and metavariables are mapped into existential variables
ProofTree2Xml: mapping from the proof-trees to a XML stream
Xmlcommand: main functions and mapping from the Coq internal data structures
            to the data structures for objects declared in the Acic module

The code compiles, but it is still very bugged (e.g. w.r.t. section variables
and proof-tree exporting). Moreover it is also very incomplete (e.g. just
a few amount of the proof-tree related informations is exported)


git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/coq/branches/mowgli@2910 85f007b7-540e-0410-9357-904b9bb8a0f7