Due to the new directory organization (i.e. no more directory for sections), reference to variables (that still have a sectioning notion) have been changed from relative ascending "URI"s (i.e. name + number of directory to skip going up) to relative descending URIs (real relative URIs, this time). A similar thing has been done for parameters, that are now just simple lists of relative URIs. (According to my experience, the old compact representation was not really relevant for avoiding space waste.
Moreover, the links from the inner-types and proof-trees files to their
corresponding logical files have been fixed.

The library should now be fully linked, without any more dandling or
simply incorrect or not meaningful pointers.

Note: for the above sentence to be true, constants with positive strength
must be removed from the system, as decided at the Eindhoven meeting.


git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/coq/branches/mowgli@2984 85f007b7-540e-0410-9357-904b9bb8a0f7