Corrected a typo in doc/refman/Setoid.tex.
Redefined the \index command in doc/refman/headers.tex only for Hevea.
Now all sectioning commands (from \part to \subparagraph) store the
value of their counter in the command \@indexlabel. It is this command
that is used inside the new \index. Thus, the index shows only the
the most recent sectioning command, but not \index, \theorem, etc.


git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/coq/trunk@9745 85f007b7-540e-0410-9357-904b9bb8a0f7