Extraction now checks that the required libraries are indeed loaded (bug #1144)
If a library is Require'd from inside an "opaque" Module (e.g. a module with 
an interface given by a ":"), then the required library is not loaded anymore 
after closing this module. We add an error in this situation, asking the user 
to manually do a Require before performing the Extraction. 



git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/coq/trunk@10208 85f007b7-540e-0410-9357-904b9bb8a0f7