Fix the number parsing/printing for BigN/BigZ/BigQ
* In order to get via a "SearchAbout bigZ" all the generic lemmas that are 
now available, I moved bigZ to a notation for BigZ.t instead of a definition.
But this notation was interacting badly with the number parsing. 

* Moreover, I also discovered that printing was broken as well, due to a 
reference to obsolete name Basic_type (now DoubleType) that I forgot to 
adapt in g_intsyntax.

* Last but not least, there was also a bug not due to myself that was probably
preventing parsing/printing _very_ big numbers (the ones for which generic 
constructor BigN.Nn is used): this Nn is the 7-th constructor, not the 13-th.
In addition, the first argument to Nn was supposed to be automatically inferred, 
this was not working, so we build it now explicitely (very easy).




git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/coq/trunk@11087 85f007b7-540e-0410-9357-904b9bb8a0f7