Restores behaviour of v8.1 for unification problems which fail: previous version had an (at least) exponential behaviour on terms "easily" detected as not being comparable. We mainly cut the recursive exploration when the kernel comparison negative answer can and should be trusted.
This fixes the problems encountered in the ssreflect contrib : all the
previous (v8.1) tactics which are commented as (* was ... *) were
killing coq up to the previous version of unification.ml in the v8.2
branch. Now they are immediate again, like they used to be in v8.1.

Unfortunately this does not reduce signlficantly the global
compilation time, at least for std lib and ssreflect contrib.



git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/coq/branches/v8.2@11585 85f007b7-540e-0410-9357-904b9bb8a0f7