Revert to old-style Zdiv and ZOdiv (no use of Numbers/Abstract/ZDiv*.v)
It seems that making ZArith depend on ZBinary (the instantiation of
Numbers by Z) induce a important time and vo-size penalty on several
contribs (around +30% in some case). Since the benefit for the moment
is anyway not tremedous (Zdiv and ZOdiv with shorter proof scripts),
we backtrack this change for 8.3, and will try to solve the issue later.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/coq/branches/v8.3@12750 85f007b7-540e-0410-9357-904b9bb8a0f7