Improved check is_open_canonical_projection
The check looks for 1 canonical projection applied to a meta/evar.
This fails to deal with telescopes that generate unification problems
containing something like "(pi_1 (pi_2 ?))" that is indeed a "stuck"
canonical projection but not of the form recognized by the previous
implementation. The same holds when pi_2 is a general function not
producing a constructor.

This patch checks if the argument of the canonical projection weak head
reduces to a constructor calling whd_betadeltaiota.