Completing r14448 and thus continuing r14407 (using Cast to propagate information coming from tactics on how to solve cst/cst critical pairs in the kernel conversion machine).
In r14448, extra Cast's were removed from kernel type-checker but
(erroneously) not from the terms actually registered in the
environment.

The current commit completes the work by registering the term output
by the type-checker and not the original term. Note that this needs to
move hconsing from before to after typing.

On the Coq library, propagating Cast (without keeping them on disk)
induces a stable 1% speedup (Xeon w3540). Having hcons before or after
type-checking makes no difference. It remains to test on user contribs
whether the current commit compensates the slow down and vo size
increasing coming with the improvement made to Qed in r14407.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/coq/trunk@14492 85f007b7-540e-0410-9357-904b9bb8a0f7