Checker: finish fixing #2609 by removing env (and hence some add_module) in strengthen
This is actually a backport of the checker part of commit 14134.
Without that, an env is maintained during strengthen, triggering
some add_module that might actually already have been done elsewhere,
hence an anomaly now that we strictly check that an env can only grow,
but not be modified in place.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/coq/branches/v8.3@14788 85f007b7-540e-0410-9357-904b9bb8a0f7