Fix failure to compile doc/stdlib/Library.tex.
Coqdoc converts the utf8 symbol lambda (that appears in Utf8_core.v) to
itself when outputting utf8. Since Library.tex uses utf8x as the input
encoding, it gets translated to \textlambda. This command is defined by
both the LGR font encoding and the tipa package, and only by them. So the
build fails.

There are several solutions:
1. \usepackage[mathletters]{ucs}
2. \usepackage[T1,LGR]{fontenc}
3. \usepackage{tipa}
4. modify coqdoc

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/coq/trunk@15807 85f007b7-540e-0410-9357-904b9bb8a0f7