the new .dir-locals.el thanks to Pierre Courtieu
As it says :
Set default directory to coq root ONLY IF variable
coq-project-find-file is non nil. This should remain a
user preference and not be set by default. This setting
is redundant with compile-command above as M-x compile
always CD's to default directory. To enable it add this
to your emacs config: (setq coq-project-find-file t)

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/coq/trunk@15924 85f007b7-540e-0410-9357-904b9bb8a0f7