Makefile: stop using ln -s in win32
In Win32, cygwin provides an emulation of ln -s, but this emulation
won't work outside of cygwin shell (i.e. typically in a Sys.command).
So we just forget about it, and do a simple copy.

This is just a workaround, the long-term solution is already in trunk:
do not produce coqtop.opt(.exe) anymore, but only coqtop(.exe).

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/coq/branches/v8.4@15936 85f007b7-540e-0410-9357-904b9bb8a0f7