Fixed a bug in the algorithm trying to elaborate a "match" return predicate.
This occurred in a second choice strategy which is anyway probably not
used in general, but the bug showed up in error messages since these
messages reports on the second choice when the first choice has failed
because of user-side typing error.

Anyway, having two strategies for building a return predicate will
probably eventually disappear with the increase of the strength of
unification.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/coq/trunk@16078 85f007b7-540e-0410-9357-904b9bb8a0f7