Mod_subst: an extra assert
It seems that the Equiv delta_hint in a resolver have a particular shape:
a kn can apparently only be shifted to another kn with same label
this way. We validate this fact via an assert, since this isn't
obvious (due to recursive calls in Mod_subst and Modops), and since
this implies the important fact that user and canonical parts
of kernel pairs have necessarily the same label (and section dirpath).

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/coq/trunk@16219 85f007b7-540e-0410-9357-904b9bb8a0f7