Backport of interface.mli and serialize.ml reworked to avoid copy/paste of types
With this commit the types involved in the protocol between Coq and
Coqide are written once and for all in interface.mli

The benefit of this commit is that changing the protocol is easier:
one changes the types in interface.mli and lets ocamlc spot all the
places that needs to be modified.
This is a necessity if one plays with the protocol very often, like
in my Paral-ITP branch.

This is revision trunk@16438

Conflicts:
	ide/coq.ml
	toplevel/ide_intf.ml
	toplevel/ide_intf.mli
	toplevel/ide_slave.ml